pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "cup-be"
version = "0.2.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=4.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "pytest>=9.0.2",
    "ruff>=0.9.1",
]
